requirements.txt
streamlit
pandas
scikit-learn
numpy
Pillow
matplotlib
seaborn

